--- app/static/css/style.css
@@ -1,6 +1,0 @@
-h1 {
-    border: 2px #eee solid;
-    color: brown;
-    text-align: center;
-    padding: 10px;
-}
