+++ src/covidVisualizationTool/app/static/css/style.css
@@ -1,0 +1,6 @@
+h1 {
+    border: 2px #eee solid;
+    color: brown;
+    text-align: center;
+    padding: 10px;
+}
