+++ src/covidVisualizationTool/app/templates/index.html
@@ -1,0 +1,5 @@
+{% extends 'base.html' %}
+
+{% block content %}
+    <h1>{% block title %} Welcome to the Covid Visualization Tool {% endblock %}</h1>
+{% endblock %}
