+++ tests/import_data/test_import.py
@@ -1,0 +1,16 @@
+from covidVisualizationTool.imports.importData import ImportData
+from common_utilities import get_file_from_path
+from covidVisualizationTool.queries.queryData import QueryData
+
+
+def test_load_tables(test_db_location):
+    test_csv_file = get_file_from_path('../../docs/owid-covid-data.csv', __file__)
+    LD = ImportData(db_location=test_db_location, csv_file=test_csv_file)
+
+    assert LD.load_tables()
+
+
+def test_simple_query(test_db_location):
+    test = QueryData(test_db_location).get_all_data()
+
+    # assert test == 457
